Improved the ngram server support
If a ngram server was launched with  `start_ngram_server.sh`, it works as before
Else the `Decoder` class will launch one for its lifespan. At it's creation the ngram server will be opened, and at its deletion it will be closed.

`start_ngram_server.sh` outputs the PID of the ngram server.

It is possible to specify the port on which the server will be launch.